be leberal about where pc-autoinstall.conf and pc-sysinstall.conf are placed and read them from /tmp when running pc-sysinstall from TrueOSStart.sh
SetupInstall.sh probed for pc-autoinstall.conf in /boot and copied it to
/tmp. Then it checks for the pc-sysinstall.conf file at the path specified in
pc-autoinstall.conf and also copy it to /tmp or aborts with an error

TrueOSStart.sh would assume to find pc-autoinstall.conf at the
filesystem root, not in /tmp where it was put by SetupInstall. Therefore
it is possible that pc-sysinstall.conf might no longer be accessible due
to mounts/overlays during boot and initial setup of the installer
environment.

With this fix SetupInstall.sh copies pc-autoinstall.conf from either "/" or
"/boot" to "/tmp" as well as the pc-sysinstall.conf (if present).
The path to pc-sysinstall.conf within pc-autoinstall.conf is also modified
to read it from /tmp.